(PC-17795) refactor: remove InlineSection
